+++ desktop/applis/apps/core/games/pokemoncore/src/main/java/aiki/map/ChangeNbPlaceFieldActTileMiniMap.java
@@ -17,7 +17,7 @@ public final class ChangeNbPlaceFieldActTileMiniMap implements ChangeNbPlaceFiel
 
     @Override
     public void place(int _p) {
-        coords.setPlace((short) _p);
+        coords.setPlace(_p);
     }
 
     @Override
